pyproject.toml
[project]
name = "mxbai-vs-bge-rerank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.4.1",
    "hdrhistogram>=0.10.3",
    "mxbai-rerank>=0.1.3",
    "sentence-transformers>=3.4.1",
]

